+++ src/References/Interval.php
@@ -12,8 +12,13 @@ interface Interval
     /**
      * Supported intervals.
      */
-    public const
-        DAY   = 'Day',
-        WEEK  = 'Week',
-        MONTH = 'Month';
+    public const DAY = 'Day';
+    public const WEEK = 'Week';
+    public const MONTH = 'Month';
+
+    public const INTERVALS_LIST = [
+        self::DAY,
+        self::WEEK,
+        self::MONTH,
+    ];
 }