+++ src/Requests/Payments/SBP/SBPQrImageRequestBuilder.php
@@ -6,24 +6,9 @@ namespace AvtoDev\CloudPayments\Requests\Payments\SBP;
 
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
-use AvtoDev\CloudPayments\Requests\Traits\HasReceipt;
-use AvtoDev\CloudPayments\Requests\AbstractRequestBuilder;
-use AvtoDev\CloudPayments\Requests\Traits\PaymentRequestTrait;
 
-class SBPQrImageRequestBuilder extends AbstractRequestBuilder
+class SBPQrImageRequestBuilder extends AbstractSBPPaymentRequestBuilder
 {
-    use PaymentRequestTrait, HasReceipt;
-
-    /**
-     * @inheritdoc
-     */
-    protected function getRequestPayload(): array
-    {
-        $this->setJsonData(\array_merge($this->json_data ?? [], $this->getReceiptData()));
-
-        return $this->getCommonPaymentParams();
-    }
-
     /**
      * @inheritdoc
      */