+++ tests/Unit/Requests/Payments/SBP/SBPQrLinkRequestBuilderTest.php
@@ -6,13 +6,15 @@ namespace AvtoDev\Tests\Unit\Requests\Payments\SBP;
 
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
-use AvtoDev\Tests\Unit\Requests\AbstractRequestBuilderTestCase;
+use PHPUnit\Framework\Attributes\CoversClass;
 use AvtoDev\CloudPayments\Requests\Payments\SBP\SBPQrLinkRequestBuilder;
+use AvtoDev\CloudPayments\Requests\Payments\SBP\AbstractSBPPaymentRequestBuilder;
 
-/**
- * @covers \AvtoDev\CloudPayments\Requests\Payments\SBP\SBPQrLinkRequestBuilder
- */
-class SBPQrLinkRequestBuilderTest extends AbstractRequestBuilderTestCase
+#[
+    CoversClass(AbstractSBPPaymentRequestBuilder::class),
+    CoversClass(SBPQrLinkRequestBuilder::class),
+]
+class SBPQrLinkRequestBuilderTest extends AbstractSBPPaymentRequestBuilderTestCase
 {
     /**
      * @inheritdoc